refactor: change path and event plugins to follow the same convention as `window`
